feat: Refactor Enable And Disable Module Commands
