+++ cmd/kyma/alpha/alpha.go
@@ -1,10 +1,10 @@
 package alpha
 
 import (
+	"github.com/kyma-project/cli/cmd/kyma/alpha/add"
 	"github.com/kyma-project/cli/cmd/kyma/alpha/create"
+	"github.com/kyma-project/cli/cmd/kyma/alpha/delete"
 	"github.com/kyma-project/cli/cmd/kyma/alpha/deploy"
-	"github.com/kyma-project/cli/cmd/kyma/alpha/disable"
-	"github.com/kyma-project/cli/cmd/kyma/alpha/enable"
 	"github.com/kyma-project/cli/cmd/kyma/alpha/list"
 	"github.com/kyma-project/cli/cmd/kyma/alpha/sign"
 	"github.com/kyma-project/cli/cmd/kyma/alpha/verify"
@@ -23,8 +23,8 @@ func NewCmd(o *cli.Options) *cobra.Command {
 
 	cmd.AddCommand(create.NewCmd(o))
 	cmd.AddCommand(list.NewCmd(o))
-	cmd.AddCommand(enable.NewCmd(o))
-	cmd.AddCommand(disable.NewCmd(o))
+	cmd.AddCommand(add.NewCmd(o))
+	cmd.AddCommand(delete.NewCmd(o))
 	cmd.AddCommand(deploy.NewCmd(deploy.NewOptions(o)))
 	cmd.AddCommand(sign.NewCmd(o))
 	cmd.AddCommand(verify.NewCmd(o))
