+++ docs/gen-docs/kyma_alpha_delete_module.md
@@ -1,43 +1,38 @@
 ---
-title: kyma alpha enable module
+title: kyma alpha delete module
 ---
 
-Enables a module in the cluster or in the given Kyma resource.
+Disables a module in the cluster or in the given Kyma resource.
 
 ## Synopsis
 
-Use this command to enable Kyma modules available in the cluster.
+Use this command to disable active Kyma modules in the cluster.
 
 ### Detailed description
 
 For more information on Kyma modules, see the 'create module' command.
 
-This command enables an available module in the cluster. 
-A module is available when it is released with a ModuleTemplate. The ModuleTemplate is used for instantiating the module with proper default configuration.
+This command disables an active module in the cluster.
 
 
 ```bash
-kyma alpha enable module [name] [flags]
+kyma alpha delete module [name] [flags]
 ```
 
 ## Examples
-
+* Delete "my-module" from the "alpha" channel in "default-kyma" in "kyma-system" Namespace
 ```bash
-
-Enable "my-module" from "alpha" channel in "default-kyma" in "kyma-system" Namespace
-		kyma alpha enable module my-module -c alpha -n kyma-system -k default-kyma
-
+kyma alpha delete module my-module -c alpha -n kyma-system -k default-kyma
 ```
 
 ## Flags
 
 ```bash
-  -c, --channel string     Module's channel to enable.
+  -c, --channel string     Module's channel to use.
   -f, --force-conflicts    Forces the patching of Kyma spec modules in case their managed field was edited by a source other than Kyma CLI.
   -k, --kyma-name string   Kyma resource to use. If empty, 'default-kyma' is used. (default "default-kyma")
   -n, --namespace string   Kyma Namespace to use. If empty, the default 'kyma-system' Namespace is used. (default "kyma-system")
-  -p, --policy string      Determines how the module is managed. Use 'CreateAndDelete' to install the default values provided in the module template or 'Ignore' to ignore them. (default "CreateAndDelete")
-  -t, --timeout duration   Maximum time for the operation to enable a module. (default 1m0s)
+  -t, --timeout duration   Maximum time for the operation to disable a module. (default 1m0s)
   -w, --wait               Wait until the given Kyma resource is ready.
 ```
 
@@ -53,5 +48,5 @@ Enable "my-module" from "alpha" channel in "default-kyma" in "kyma-system" Names
 
 ## See also
 
-* [kyma alpha enable](kyma_alpha_enable.md)	 - Enables a resource in the Kyma cluster.
+* [kyma alpha delete](kyma_alpha_delete.md)	 - Disables a resource in the Kyma cluster.
